+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.8] - 2025-04-08
+
+### Added
+
+- Adds support for the "<select>" and "<option>" tags using the "Goatee.Select()" and "Goatee.Option()" classes.
+- Adds an example application to demonstrate how to use Goatee.
+
 ## [0.1.7] - 2025-04-08
 
 ### Changes